fix: specmatic mcp exposes tools which can be accessed by the coding agents
